get_balance handler correct

+++ user.go
@@ -1,9 +1,9 @@
 package User_balance
 
 type User struct {
-	Id       int    `json:"-" db:"id"`
-	Name     string `json:"name" binding:"required"`
-	Username string `json:"username" binding:"required"`
-	Password string `json:"password" binding:"required"`
-	Balance  int    `json:"balance"`
+	Id       int     `json:"-" db:"id"`
+	Name     string  `json:"name" binding:"required"`
+	Username string  `json:"username" binding:"required"`
+	Password string  `json:"password" binding:"required"`
+	Balance  float32 `json:"balance"`
 }
